extracted boundary condition object logic for (re-)moving of objects into its own file. still work in progress
